requirements.txt
pandas 
numpy
PyYAML
scikit-learn==1.0.2
flask
